+++ CHANGELOG.md
@@ -2,3 +2,9 @@
 
 ### Changed
 -  Remove sudo from clearOldLogs command.
+
+### Added
+- Add logout function.
+
+### Performance
+- Optimize variable.